custom-nodes/node-red-contrib-facedetectionnode/requirements.txt
numpy
opencv-python
ultralytics
facenet-pytorch
retinaface
Pillow
